fixed nested listeners
