+++ src/scripts/login/dashboardActivator.js
@@ -13,10 +13,10 @@ const dashboardActivator = () => {
     friendActivator()
 
     //this prints the tasks section after login - MT
-    tasks.afterLogin(sessionStorage.getItem("activeUser"));
+    tasks.tasksActivator(sessionStorage.getItem("activeUser"));
 
     //This is just a placeholder until we get all the other pieces
-    document.querySelector("#header").innerHTML = `you are logged in`
+    document.querySelector("#header").innerHTML += `you are logged in`
 
     //Adds event dashboard
     eventDashboard();
